Compiles with one lexer and one parser
