+++ README.md
@@ -1,0 +1,4 @@
+1. Simpler expressions
+2. Define for calc_value_t so it can be changed in compilation process
+3. Single lexer and parser
+4. "Stairs" in ast mode — cache levels
